first draft of extenable aas view architecture
