+++ compose.yml
@@ -30,7 +30,6 @@ services:
       AAS_LIST_FEATURE_FLAG: "true"
       TRANSFER_FEATURE_FLAG: "false"
       COMPARISON_FEATURE_FLAG: "true"
-      EXPERIMENTAL_PRODUCT_VIEW_FEATURE_FLAG: "false"
       AUTHENTICATION_FEATURE_FLAG: "false"
       LOCK_TIMESERIES_PERIOD_FEATURE_FLAG: "true"
       NEXTAUTH_SECRET: ${NEXTAUTH_SECRET:-verySecureNextAuthSecret}
