+++ src/app/EnvProvider.tsx
@@ -10,7 +10,6 @@ export const initialEnvValues: EnvironmentalVariables = {
     AAS_LIST_FEATURE_FLAG: false,
     COMPARISON_FEATURE_FLAG: false,
     TRANSFER_FEATURE_FLAG: false,
-    EXPERIMENTAL_PRODUCT_VIEW_FEATURE_FLAG: false,
     KEYCLOAK_ENABLED: false,
     LOCK_TIMESERIES_PERIOD_FEATURE_FLAG: false,
     DISCOVERY_API_URL: undefined,
