+++ src/app/[locale]/list/_components/AasListTableRow.tsx
@@ -16,7 +16,6 @@ import { MultiLanguageValueOnly } from 'lib/api/basyx-v3/types';
 import { useLocale, useTranslations } from 'next-intl';
 import { encodeBase64 } from 'lib/util/Base64Util';
 import useSWR from 'swr';
-import { useEnv } from 'app/EnvProvider';
 import { RepositoryWithInfrastructure } from 'lib/services/database/InfrastructureMappedTypes';
 
 type AasTableRowProps = {
@@ -48,7 +47,6 @@ export const AasListTableRow = (props: AasTableRowProps) => {
     const notificationSpawner = useNotificationSpawner();
     const [thumbnailUrl, setThumbnailUrl] = useState<string>('');
     const t = useTranslations('pages.aasList');
-    const env = useEnv();
     const locale = useLocale();
     const { data: nameplateValues, isLoading: isNameplateValueLoading } = useSWR(
         [repository, aasListEntry.aasId],
@@ -66,7 +64,7 @@ export const AasListTableRow = (props: AasTableRowProps) => {
 
     const navigateToAas = (listEntry: ListEntityDto) => {
         const baseUrl = window.location.origin;
-        const pageToGo = env.EXPERIMENTAL_PRODUCT_VIEW_FEATURE_FLAG ? '/product' : '/viewer';
+        const pageToGo = '/viewer';
 
         // Only send repoUrl parameter if it's a repository, not a registry
         const repoUrlParam = connectionType === 'repository' && repository.url ? `?repoUrl=${repository.url}` : '';
