+++ src/app/[locale]/viewer/[base64AasId]/layout.tsx
@@ -1,14 +1,20 @@
 'use client';
 
 import { Box } from '@mui/material';
+import { PropsWithChildren } from 'react';
 import { safeBase64Decode } from 'lib/util/Base64Util';
 import { useParams, useSearchParams } from 'next/navigation';
-import { NoSearchResult } from 'components/basics/detailViewBasics/NoSearchResult';
 import { CurrentAasContextProvider } from 'components/contexts/CurrentAasContext';
+import { NoSearchResult } from 'components/basics/detailViewBasics/NoSearchResult';
 import { useShowError } from 'lib/hooks/UseShowError';
-import { ProductViewer } from '../_components/ProductViewer';
+import { AasViewSwitcher } from 'app/[locale]/viewer/_components/AasViewSwitcher';
 
-export default function Page() {
+/**
+ * Shared layout for a single AAS. Hoists {@link CurrentAasContextProvider} so
+ * AAS + submodel data is fetched once and preserved across view switches under
+ * the `[view]` segment, and renders the {@link AasViewSwitcher} above the view.
+ */
+export default function AasViewerLayout({ children }: PropsWithChildren) {
     const { showError } = useShowError();
     const params = useParams<{ base64AasId: string }>();
     const base64AasId = decodeURIComponent(params.base64AasId).replace(/=+$|[%3D]+$/, '');
@@ -16,14 +22,11 @@ export default function Page() {
     const repoUrl = encodedRepoUrl ? decodeURI(encodedRepoUrl) : undefined;
     const infrastructureName = useSearchParams().get('infrastructure') || undefined;
 
-    let aasIdDecoded: string | null = null;
+    let aasIdDecoded: string;
     try {
         aasIdDecoded = safeBase64Decode(base64AasId);
     } catch (e) {
         showError(e);
-    }
-
-    if (!aasIdDecoded) {
         return (
             <Box
                 sx={{
@@ -43,7 +46,8 @@ export default function Page() {
 
     return (
         <CurrentAasContextProvider aasId={aasIdDecoded} repoUrl={repoUrl} infrastructureName={infrastructureName}>
-            <ProductViewer />
+            <AasViewSwitcher />
+            {children}
         </CurrentAasContextProvider>
     );
 }
