+++ src/app/[locale]/viewer/[base64AasId]/page.tsx
@@ -1,43 +1,23 @@
-'use client';
-import { CurrentAasContextProvider } from 'components/contexts/CurrentAasContext';
-import { safeBase64Decode } from 'lib/util/Base64Util';
-import { useParams, useSearchParams } from 'next/navigation';
-import { AASViewer } from '../_components/AASViewer';
-import { NoSearchResult } from 'components/basics/detailViewBasics/NoSearchResult';
-import { useShowError } from 'lib/hooks/UseShowError';
-import { Box } from '@mui/material';
+import { redirect } from 'next/navigation';
+import { aasViewerConfig } from 'app/[locale]/viewer/_visualizations/viewer.config';
 
-export default function () {
-    const { showError } = useShowError();
-    const params = useParams<{ base64AasId: string }>();
-    const base64AasId = decodeURIComponent(params.base64AasId).replace(/=+$|[%3D]+$/, '');
-    const encodedRepoUrl = useSearchParams().get('repoUrl');
-    const repoUrl = encodedRepoUrl ? decodeURI(encodedRepoUrl) : undefined;
-    const infrastructureName = useSearchParams().get('infrastructure') || undefined;
-    try {
-        const aasIdDecoded = safeBase64Decode(base64AasId);
+type ViewerPageProps = {
+    params: Promise<{ base64AasId: string }>;
+    searchParams: Promise<Record<string, string | string[] | undefined>>;
+};
 
-        return (
-            <CurrentAasContextProvider aasId={aasIdDecoded} repoUrl={repoUrl} infrastructureName={infrastructureName}>
-                <AASViewer />
-            </CurrentAasContextProvider>
-        );
-    } catch (e) {
-        showError(e);
-        return (
-            <Box
-                sx={{
-                    display: 'flex',
-                    flexDirection: 'column',
-                    gap: '30px',
-                    alignItems: 'center',
-                    width: '100vw',
-                    marginBottom: '50px',
-                    marginTop: '20px',
-                }}
-            >
-                <NoSearchResult base64AasId={base64AasId} />
-            </Box>
-        );
-    }
+/**
+ * Bare `/viewer/<id>` entry point. Redirects to the configured default view
+ * under the `[view]` segment, preserving `repoUrl` / `infrastructure` params.
+ */
+export default async function ViewerPage({ params, searchParams }: ViewerPageProps) {
+    const { base64AasId } = await params;
+    const sp = await searchParams;
+
+    const query = new URLSearchParams();
+    if (typeof sp.repoUrl === 'string') query.set('repoUrl', sp.repoUrl);
+    if (typeof sp.infrastructure === 'string') query.set('infrastructure', sp.infrastructure);
+    const suffix = query.toString() ? `?${query.toString()}` : '';
+
+    redirect(`/viewer/${base64AasId}/${aasViewerConfig.default}${suffix}`);
 }
