+++ src/app/[locale]/viewer/_components/product/KeyFactsBox.tsx
