+++ src/app/[locale]/viewer/_components/product/ProductOverviewCard.tsx
@@ -11,13 +11,13 @@ import { useTranslations } from 'next-intl';
 import { SubmodelSemanticIdEnum } from 'lib/enums/SubmodelSemanticId.enum';
 import { findSubmodelByIdOrSemanticId, findSubmodelElementByIdShort } from 'lib/util/SubmodelResolverUtil';
 import { MobileAccordion } from 'components/basics/detailViewBasics/MobileAccordion';
-import { KeyFactsBox } from 'app/[locale]/product/_components/KeyFactsBox';
+import { KeyFactsBox } from 'app/[locale]/viewer/_components/product/KeyFactsBox';
 import { SubmodelElementSemanticIdEnum } from 'lib/enums/SubmodelElementSemanticId.enum';
 import { useProductImageUrl } from 'lib/hooks/UseProductImageUrl';
 import { useFindValueByIdShort } from 'lib/hooks/useFindValueByIdShort';
 import { ActionMenu } from './ProductActionMenu';
 import LinkIcon from '@mui/icons-material/Link';
-import { MnestixConnection } from '../../../../../prisma/generated/client';
+import { MnestixConnection } from '../../../../../../prisma/generated/client';
 import {
     AssetAdministrationShell,
     Property,
@@ -380,9 +380,11 @@ export function ProductOverviewCard(props: ProductOverviewCardProps) {
     const classificationInfo = (
         <Box sx={infoBoxStyle} data-testid="asset-data">
             {!isAccordion && (
-                <Box sx={{
-                    display: 'flex'
-                }}>
+                <Box
+                    sx={{
+                        display: 'flex',
+                    }}
+                >
                     <IconCircleWrapper sx={{ mr: 1 }}>
                         <AssetIcon fontSize="small" color="primary" />
                     </IconCircleWrapper>
@@ -422,9 +424,11 @@ export function ProductOverviewCard(props: ProductOverviewCardProps) {
                             sx={{ height: '300px', maxWidth: '300px', width: '100%' }}
                             data-testid="aas-loading-skeleton"
                         ></Skeleton>
-                        <Box sx={{
-                            width: '100%'
-                        }}>
+                        <Box
+                            sx={{
+                                width: '100%',
+                            }}
+                        >
                             {isAccordion ? (
                                 <Box sx={{ m: 1 }}>
                                     <Skeleton width="100%" />
@@ -500,8 +504,9 @@ export function ProductOverviewCard(props: ProductOverviewCardProps) {
                                                         sx={{
                                                             fontSize: 'small',
                                                             ml: 1,
-                                                            cursor: 'pointer'
-                                                        }} />
+                                                            cursor: 'pointer',
+                                                        }}
+                                                    />
                                                 </Tooltip>
                                             )}
                                         </Typography>
