+++ src/components/basics/listBasics/GenericAasListEntry.tsx
@@ -7,7 +7,6 @@ import { RoundedIconButton } from 'components/basics/Buttons';
 import { ArrowForward } from '@mui/icons-material';
 import { useTranslations } from 'next-intl';
 import { ImageWithFallback } from 'components/basics/StyledImageWithFallBack';
-import { useEnv } from 'app/EnvProvider';
 
 type GenericAasListTableRowProps = {
     aasListEntry: AasListEntry;
@@ -21,11 +20,10 @@ const tableBodyText = {
 
 export const GenericAasListEntry = ({ aasListEntry, ...config }: GenericAasListTableRowProps) => {
     const navigate = useRouter();
-    const env = useEnv();
     const t = useTranslations('pages.aasList');
 
     const navigateToAas = () => {
-        const pageToGo = env.EXPERIMENTAL_PRODUCT_VIEW_FEATURE_FLAG ? '/product' : '/viewer';
+        const pageToGo = '/viewer';
         const aasId = aasListEntry.aasId;
         const repoUrl = aasListEntry.repositoryUrl;
         const infrastructureName = aasListEntry.infrastructureName;
