+++ src/locale/en.json
@@ -219,8 +219,11 @@
     "aasViewer": {
       "actions": {
         "compareButton": "Compare",
-        "toProductView": "Product view",
         "download": "Download AAS"
+      },
+      "views": {
+        "default": "AAS View",
+        "product": "Product view"
       },
       "asset": "Asset",
       "assetAdministrationShell": "Asset Administration Shell",
@@ -339,8 +342,7 @@
         "compareButton": "Compare",
         "download": "Download AAS",
         "downloadError": "Error while downloading AAS.",
-        "downloadErrorNoRepo": "No repository found for the selected AAS.",
-        "toAasView": "AAS View"
+        "downloadErrorNoRepo": "No repository found for the selected AAS."
       },
       "classification": "Classification",
       "error": {
