+++ src/locale/es.json
@@ -219,8 +219,11 @@
     "aasViewer": {
       "actions": {
         "compareButton": "Comparar",
-        "toProductView": "Vista de producto",
         "download": "Descargar AAS"
+      },
+      "views": {
+        "default": "Vista AAS",
+        "product": "Vista de producto"
       },
       "asset": "Asset",
       "assetAdministrationShell": "Asset Administration Shell",
@@ -339,8 +342,7 @@
         "compareButton": "Comparar",
         "download": "Descargar AAS",
         "downloadError": "Error al descargar el AAS.",
-        "downloadErrorNoRepo": "No se encontró ningún repositorio para el AAS seleccionado.",
-        "toAasView": "Vista del AAS"
+        "downloadErrorNoRepo": "No se encontró ningún repositorio para el AAS seleccionado."
       },
       "classification": "Clasificación del producto",
       "error": {
