+++ wiki/Mnestix-Configuration-Settings.md
@@ -30,10 +30,9 @@ Mnestix provides the following configuration options. You can adapt the values i
 
 #### Experimental Features
 
-| Name                                     | Default value | Description                                                                                                                                                                                                                                                                                                                                                                                                                                                           |
-| ---------------------------------------- | ------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| `EXPERIMENTAL_PRODUCT_VIEW_FEATURE_FLAG` | false         | Enables or disables the experimental product view feature.                                                                                                                                                                                                                                                                                                                                                                                                            |
-| `EXPERIMENTAL_HIGHLIGHT_DATA_FLAG`       | false         | Enables highlighting of specific submodels, properties, and multilang properties rendered with the default/fallback visualization. When enabled, elements can be visually emphasized using a `HighlightColor` qualifier. Example qualifier configuration: `{"kind": "ValueQualifier", "type": "HighlightColor", "value": "rgb(255,0,0)", "valueType": "xs:string"}`. For visual examples, see [PR #491](https://github.com/eclipse-mnestix/mnestix-browser/pull/491). |
+| Name                               | Default value | Description                                                                                                                                                                                                                                                                                                                                                                                                                                                           |
+| ---------------------------------- | ------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| `EXPERIMENTAL_HIGHLIGHT_DATA_FLAG` | false         | Enables highlighting of specific submodels, properties, and multilang properties rendered with the default/fallback visualization. When enabled, elements can be visually emphasized using a `HighlightColor` qualifier. Example qualifier configuration: `{"kind": "ValueQualifier", "type": "HighlightColor", "value": "rgb(255,0,0)", "valueType": "xs:string"}`. For visual examples, see [PR #491](https://github.com/eclipse-mnestix/mnestix-browser/pull/491). |
 
 #### External Links
